S302T700-464 #time 1h 10m Modified sales report endpoint to break up the report by days. Fixed sql queries and dto. ?implement ?fix ?pair[ojc31,pwt25]

+++ backend/src/main/java/com/seng302/wasteless/service/PurchasedListingService.java
@@ -63,7 +63,7 @@ public class PurchasedListingService {
      * @param endDate       The end date for the date range. Format yyyy-MM-dd
      * @return The count of purchased listings
      */
-    public Integer totalPurchasedListingValueForBusinessInDateRange(Integer businessId, LocalDate startDate, LocalDate endDate) {
+    public Double totalPurchasedListingValueForBusinessInDateRange(Integer businessId, LocalDate startDate, LocalDate endDate) {
         return purchasedListingRepository.sumPriceByBusiness_IdAndSaleDateBetween(businessId, startDate, endDate);}
 
 }
